editor 4-29 19.30

+++ cnc-app/Domain/Views/DeviceCNC.cs
@@ -20,7 +20,7 @@ namespace APP.Domain.Views
 
         public string? Host { get; set; }
 
-        public long? Port { get; set; }
+        public int? Port { get; set; }
 
         public long? ReadTimeOut { get; set; }
